Documents/Materi smt 7/Algoritma Deep Learning/tugas-deep-learning/requirements.txt
Flask==2.2.5
gunicorn==21.2.0

# YOLO & CV
ultralytics==8.2.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==9.5.0

# ML lain
tensorflow-cpu==2.12.0
pandas==2.0.3
matplotlib==3.7.1
yfinance==0.2.28
scikit-learn==1.2.2
